QAPPBuilder/requirements.txt
# ########################################################################## #
# NOTE: Do not install from this file. This file is used to check for
# dependency security vulnerabilities via Dependabot.
# To install pip packages, use the requirements_essential.txt file.
# ########################################################################## #
asgiref==3.8.1
astroid==3.3.9
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
coverage==7.8.0
cryptography==44.0.2
defusedxml==0.7.1
dictor==0.1.12
dill==0.3.9
Django==5.1.6
django-docs==0.3.3
django-extensions==3.2.3
django-simple-history==3.8.0
djangorestframework==3.16.0
elementpath==4.8.0
grafana-django-saml2-auth==3.19.0
idna==3.10
isort==6.0.1
lxml==5.3.1
mccabe==0.7.0
pillow==11.1.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pydocstyle==6.3.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pysaml2==7.5.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
reportlab==4.3.1
requests==2.32.3
six==1.17.0
snowballstemmer==2.2.0
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.13.1
urllib3==2.3.0
xmlschema==2.5.1
